Refactor opening review workflow: replace queue with Celery tasks and integrate MongoDB for job management
