+++ docker/docker-compose.yml
@@ -58,7 +58,7 @@ services:
         condition: service_healthy
     networks:
       - pgnseek_net
-    command: celery -A app.tasks.celery_app worker --loglevel=info
+    command: celery -A app.review.tasks.celery_app worker --loglevel=info
 
   frontend:
     build:
